Refactor: Use u32 for key length encoding
Reduces the size of stored metadata by changing how it encodes the
length of keys in various structures within `model_loader.rs`.
Previously, key lengths were encoded using a 64-bit integer (`u64`),
which was unnecessarily large given that practical key lengths are far
shorter than $2^{32}$.

This change updates all instances to use a 32-bit integer (`u32`) for
encoding key lengths.

+++ llm-runner/src/model_loader.rs
@@ -304,7 +304,7 @@ mod tests {
         buf.extend_from_slice(&(kv_pairs.len() as u64).to_le_bytes());
         for kv in &kv_pairs {
             let key_bytes = kv.key.as_bytes();
-            buf.extend_from_slice(&(key_bytes.len() as u64).to_le_bytes());
+            buf.extend_from_slice(&(key_bytes.len() as u32).to_le_bytes());
             buf.extend_from_slice(key_bytes);
             buf.extend_from_slice(&kv.value_type.to_u32().to_le_bytes());
             write_kv_value(&mut buf, &kv.value);
@@ -424,7 +424,7 @@ mod tests {
         buf.extend_from_slice(&(kv_pairs.len() as u64).to_le_bytes());
         for kv in &kv_pairs {
             let key_bytes = kv.key.as_bytes();
-            buf.extend_from_slice(&(key_bytes.len() as u64).to_le_bytes());
+            buf.extend_from_slice(&(key_bytes.len() as u32).to_le_bytes());
             buf.extend_from_slice(key_bytes);
             buf.extend_from_slice(&kv.value_type.to_u32().to_le_bytes());
             write_kv_value(&mut buf, &kv.value);
@@ -475,7 +475,7 @@ mod tests {
         buf.extend_from_slice(&(kv_pairs.len() as u64).to_le_bytes());
         for kv in &kv_pairs {
             let key_bytes = kv.key.as_bytes();
-            buf.extend_from_slice(&(key_bytes.len() as u64).to_le_bytes());
+            buf.extend_from_slice(&(key_bytes.len() as u32).to_le_bytes());
             buf.extend_from_slice(key_bytes);
             buf.extend_from_slice(&kv.value_type.to_u32().to_le_bytes());
             write_kv_value(&mut buf, &kv.value);
@@ -526,7 +526,7 @@ mod tests {
         buf.extend_from_slice(&(kv_pairs.len() as u64).to_le_bytes());
         for kv in &kv_pairs {
             let key_bytes = kv.key.as_bytes();
-            buf.extend_from_slice(&(key_bytes.len() as u64).to_le_bytes());
+            buf.extend_from_slice(&(key_bytes.len() as u32).to_le_bytes());
             buf.extend_from_slice(key_bytes);
             buf.extend_from_slice(&kv.value_type.to_u32().to_le_bytes());
             write_kv_value(&mut buf, &kv.value);
@@ -577,7 +577,7 @@ mod tests {
         buf.extend_from_slice(&(kv_pairs.len() as u64).to_le_bytes());
         for kv in &kv_pairs {
             let key_bytes = kv.key.as_bytes();
-            buf.extend_from_slice(&(key_bytes.len() as u64).to_le_bytes());
+            buf.extend_from_slice(&(key_bytes.len() as u32).to_le_bytes());
             buf.extend_from_slice(key_bytes);
             buf.extend_from_slice(&kv.value_type.to_u32().to_le_bytes());
             write_kv_value(&mut buf, &kv.value);
@@ -628,7 +628,7 @@ mod tests {
         buf.extend_from_slice(&(kv_pairs.len() as u64).to_le_bytes());
         for kv in &kv_pairs {
             let key_bytes = kv.key.as_bytes();
-            buf.extend_from_slice(&(key_bytes.len() as u64).to_le_bytes());
+            buf.extend_from_slice(&(key_bytes.len() as u32).to_le_bytes());
             buf.extend_from_slice(key_bytes);
             buf.extend_from_slice(&kv.value_type.to_u32().to_le_bytes());
             write_kv_value(&mut buf, &kv.value);
@@ -676,7 +676,7 @@ mod tests {
         buf.extend_from_slice(&(kv_pairs.len() as u64).to_le_bytes());
         for kv in &kv_pairs {
             let key_bytes = kv.key.as_bytes();
-            buf.extend_from_slice(&(key_bytes.len() as u64).to_le_bytes());
+            buf.extend_from_slice(&(key_bytes.len() as u32).to_le_bytes());
             buf.extend_from_slice(key_bytes);
             buf.extend_from_slice(&kv.value_type.to_u32().to_le_bytes());
             write_kv_value(&mut buf, &kv.value);
@@ -745,7 +745,7 @@ mod tests {
         buf.extend_from_slice(&(kv_pairs.len() as u64).to_le_bytes());
         for kv in &kv_pairs {
             let key_bytes = kv.key.as_bytes();
-            buf.extend_from_slice(&(key_bytes.len() as u64).to_le_bytes());
+            buf.extend_from_slice(&(key_bytes.len() as u32).to_le_bytes());
             buf.extend_from_slice(key_bytes);
             buf.extend_from_slice(&kv.value_type.to_u32().to_le_bytes());
             write_kv_value(&mut buf, &kv.value);
@@ -798,7 +798,7 @@ mod tests {
         buf.extend_from_slice(&(kv_pairs.len() as u64).to_le_bytes());
         for kv in &kv_pairs {
             let key_bytes = kv.key.as_bytes();
-            buf.extend_from_slice(&(key_bytes.len() as u64).to_le_bytes());
+            buf.extend_from_slice(&(key_bytes.len() as u32).to_le_bytes());
             buf.extend_from_slice(key_bytes);
             buf.extend_from_slice(&kv.value_type.to_u32().to_le_bytes());
             write_kv_value(&mut buf, &kv.value);
@@ -846,7 +846,7 @@ mod tests {
         buf.extend_from_slice(&(kv_pairs.len() as u64).to_le_bytes());
         for kv in &kv_pairs {
             let key_bytes = kv.key.as_bytes();
-            buf.extend_from_slice(&(key_bytes.len() as u64).to_le_bytes());
+            buf.extend_from_slice(&(key_bytes.len() as u32).to_le_bytes());
             buf.extend_from_slice(key_bytes);
             buf.extend_from_slice(&kv.value_type.to_u32().to_le_bytes());
             write_kv_value(&mut buf, &kv.value);
@@ -891,7 +891,7 @@ mod tests {
         buf.extend_from_slice(&(kv_pairs.len() as u64).to_le_bytes());
         for kv in &kv_pairs {
             let key_bytes = kv.key.as_bytes();
-            buf.extend_from_slice(&(key_bytes.len() as u64).to_le_bytes());
+            buf.extend_from_slice(&(key_bytes.len() as u32).to_le_bytes());
             buf.extend_from_slice(key_bytes);
             buf.extend_from_slice(&kv.value_type.to_u32().to_le_bytes());
             write_kv_value(&mut buf, &kv.value);
@@ -952,7 +952,7 @@ mod tests {
         buf.extend_from_slice(&(kv_pairs.len() as u64).to_le_bytes());
         for kv in &kv_pairs {
             let key_bytes = kv.key.as_bytes();
-            buf.extend_from_slice(&(key_bytes.len() as u64).to_le_bytes());
+            buf.extend_from_slice(&(key_bytes.len() as u32).to_le_bytes());
             buf.extend_from_slice(key_bytes);
             buf.extend_from_slice(&kv.value_type.to_u32().to_le_bytes());
             write_kv_value(&mut buf, &kv.value);
